Split out transaction-fee types
